requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
djangorestframework==3.12.2
mysqlclient==2.0.1
Pillow==8.0.1
pycryptodome==3.9.9
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
